requirements.txt
zc.lockfile
cherrypy
githubpy